% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots a visualization of the accidents in the given state.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer or object coercible into an integer, raises an error if state number not found in data}

\item{year}{An integer or object coercible into an integer}
}
\description{
For each state, the accident data is plotted as a point by it's latitude and longitude within the state's borders.
}
\examples{
\dontrun{fars_map_state(1,'2013')}
\dontrun{fars_map_state(21,'2015')}

}
