% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{Reads a file into a dataframe.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string specifying the file from which we want to import data.}
}
\value{
dataframe A dataframe with the contents specified by the filename, if no such file exists, function will raise an error
}
\description{
Helper function for loading FARS data into a dataframe.
}
\examples{
\dontrun{fars_read('data.csv')}

}
