% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads FARS data for specified years.}
\usage{
fars_read_years(years)
}
\arguments{
\item{\code{years}}{A vector or list of strings, or objects coercible into integers representing years}
}
\value{
list A list of dataframes, raises an error if invalid year given
}
\description{
A helper function which for each given year creates a dataframe with the month and year columns from the original data
}
\examples{
\dontrun{fars_read_years(c('2013'))}
\dontrun{fars_read_years(c('2013','2014','2015'))}

}
