% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes accidents per month for each given year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{\code{years}}{A vector or list of strings, or objects coercible into integers representing years}
}
\value{
dataframe A dataframe with a column for each year
}
\description{
Creates a dataframe using the FARS data with a column for each year and a row for each month with the number of accidents in that month.
}
\examples{
\dontrun{fars_summarize_years(c('2013'))}
\dontrun{fars_summarize_years(c('2013','2014','2015'))}

}
