% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{Returns a valid filename from a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or object coercible into an integer.}
}
\value{
vector A character vector to be used as a filename
}
\description{
Helper function that generates a filename string for loading in the FARS data.
}
\examples{
\dontrun{make_filename('2013')}

}
